Refactor and adjust styling
