+++ client/src/components/SearchSubmitForm.js
@@ -1,16 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
-import SearchInput from './SearchInput';
+import SelectTVShow from './SelectTVShow';
 import SubmitButton from './SubmitButton';
-import useDebounce from '../hooks/useDebounce';
-import { searchShows } from '../api/shows';
 import { addMultipleToWatchedList, addMultipleToWatchList } from '../api/lists';
 import useAuth from '../contexts/auth/useAuth';
 import refetchQueries from '../utils/refetchQueries';
 import { useMutation } from 'react-query';
 
-const Container = styled.div`
+const Container = styled.form`
   display: flex;
   flex-flow: row wrap;
   justify-content: space-around;
@@ -51,14 +49,9 @@ const ButtonContainer = styled.div`
   }
 `;
 
+const searchInputNames = ['show1', 'show2', 'show3', 'show4'];
 function SearchSubmitForm({ tab, textvariation }) {
-  const [searchResults, setSearchResults] = useState([]);
-  const [isSearching, setIsSearching] = useState(false);
-  const [error, setError] = useState('');
-  const [selectedInputs, setSelectedInputs] = useState([]);
-  const [selectedShowsInputs, setSelectedShowsInputs] = useState({});
   const [selectedShows, setSelectedShows] = useState({});
-  const [focused, setFocused] = useState('');
   const { authenticatedUser } = useAuth();
   const userId = authenticatedUser.userId;
 
@@ -73,90 +66,39 @@ function SearchSubmitForm({ tab, textvariation }) {
     },
   });
 
-  const debouncedValue = useDebounce(selectedShowsInputs[focused], 400);
-
-  useEffect(() => {
-    if (debouncedValue) {
-      setIsSearching(true);
-      searchShows(debouncedValue)
-        .then((results) => {
-          setIsSearching(false);
-          setSearchResults(results);
-        })
-        .catch((error) => setError(error.message));
-    } else {
-      setIsSearching(false);
-      setSearchResults([]);
-    }
-  }, [debouncedValue]);
-
-  async function handleSelect(searchResult, name) {
-    setSelectedShowsInputs({
-      ...selectedShowsInputs,
-      [name]: [`${searchResult.title} (${searchResult.airYear})`],
-    });
+  async function handleSelect(searchInputName, searchResult) {
     setSelectedShows({
       ...selectedShows,
-      [name]: searchResult.id,
-    });
-    setSelectedInputs([...selectedInputs, name]);
-    setSearchResults([]);
-  }
-
-  function handleChange(event) {
-    const value = event.target.value;
-    setFocused(event.target.name);
-    if (value === '') {
-      setSelectedShowsInputs({
-        ...selectedShowsInputs,
-        [event.target.name]: null,
-      });
-      setSelectedShows({
-        ...selectedShows,
-        [event.target.name]: null,
-      });
-      const index = selectedInputs.indexOf(event.target.name);
-      if (index > -1) {
-        selectedInputs.splice(index, 1);
-        setSelectedInputs(selectedInputs);
-      }
-    }
-    setSelectedShowsInputs({
-      ...selectedShowsInputs,
-      [event.target.name]: value,
+      [searchInputName]: searchResult,
     });
   }
 
   async function handleSubmit() {
-    const selection = Object.values(selectedShows);
+    const showIds = Object.values(selectedShows).map(
+      (selectedShow) => selectedShow.id
+    );
+    console.log(showIds);
     if (tab === 'towatch') {
-      await addMultipleToWatch({ userId, selection });
+      await addMultipleToWatch({ userId, showIds });
     } else if (tab === 'watched') {
-      await addMultipleToWatched({ userId, selection });
+      await addMultipleToWatched({ userId, showIds });
     }
   }
-
-  const searchInputs = ['show1', 'show2', 'show3', 'show4'];
 
   return (
     <Container>
       <IntroductionText>
-        {`To see recommendations based on your personal viewing profile please
-        choose up to four shows you ${textvariation}`}
+        To see recommendations based on your personal viewing profile please
+        choose up to four shows you ${textvariation}
       </IntroductionText>
       <InputContainer>
-        {searchInputs.map((searchInput) => (
-          <SearchInput
-            value={selectedShowsInputs[searchInput]}
-            key={searchInput}
-            focused={focused}
-            name={searchInput}
-            searchResults={searchResults}
-            error={error}
-            isSearching={isSearching}
-            onSelect={handleSelect}
-            onChange={handleChange}
-            selectedInputs={selectedInputs}
+        {searchInputNames.map((searchInputName) => (
+          <SelectTVShow
+            key={searchInputName}
+            value={selectedShows[searchInputName]}
+            onSelect={(searchResult) =>
+              handleSelect(searchInputName, searchResult)
+            }
           />
         ))}
       </InputContainer>
